refactor: Reorganize from_iter and dispatch to collect_ca when possible (#15904)
