+++ crates/polars-core/src/chunked_array/from_iterator_par.rs
@@ -1,347 +1,15 @@
 //! Implementations of upstream traits for [`ChunkedArray<T>`]
-use std::borrow::{Borrow, Cow};
 use std::collections::LinkedList;
-use std::marker::PhantomData;
 
 use arrow::bitmap::{Bitmap, MutableBitmap};
 use polars_utils::sync::SyncPtr;
 use rayon::prelude::*;
 
-use crate::chunked_array::builder::{
-    get_list_builder, AnonymousListBuilder, AnonymousOwnedListBuilder,
-};
-#[cfg(feature = "dtype-array")]
-use crate::chunked_array::builder::{AnonymousOwnedFixedSizeListBuilder, FixedSizeListBuilder};
-#[cfg(feature = "object")]
-use crate::chunked_array::object::builder::get_object_type;
-#[cfg(feature = "object")]
-use crate::chunked_array::object::ObjectArray;
+use super::from_iterator::PolarsAsRef;
+use crate::chunked_array::builder::get_list_builder;
 use crate::prelude::*;
 use crate::utils::flatten::flatten_par;
-use crate::utils::{get_iter_capacity, NoNull};
-
-impl<T: PolarsDataType> Default for ChunkedArray<T> {
-    fn default() -> Self {
-        ChunkedArray {
-            field: Arc::new(Field::new("default", DataType::Null)),
-            chunks: Default::default(),
-            phantom: PhantomData,
-            bit_settings: Default::default(),
-            length: 0,
-            null_count: 0,
-        }
-    }
-}
-
-/// FromIterator trait
-impl<T> FromIterator<Option<T::Native>> for ChunkedArray<T>
-where
-    T: PolarsNumericType,
-{
-    fn from_iter<I: IntoIterator<Item = Option<T::Native>>>(iter: I) -> Self {
-        // TODO: eliminate this FromIterator implementation entirely.
-        iter.into_iter().collect_ca("")
-    }
-}
-
-// NoNull is only a wrapper needed for specialization
-impl<T> FromIterator<T::Native> for NoNull<ChunkedArray<T>>
-where
-    T: PolarsNumericType,
-{
-    // We use Vec because it is way faster than Arrows builder. We can do this because we
-    // know we don't have null values.
-    fn from_iter<I: IntoIterator<Item = T::Native>>(iter: I) -> Self {
-        // 2021-02-07: aligned vec was ~2x faster than arrow collect.
-        let av = iter.into_iter().collect::<Vec<T::Native>>();
-        NoNull::new(ChunkedArray::from_vec("", av))
-    }
-}
-
-impl FromIterator<Option<bool>> for ChunkedArray<BooleanType> {
-    fn from_iter<I: IntoIterator<Item = Option<bool>>>(iter: I) -> Self {
-        BooleanArray::from_iter(iter).into()
-    }
-}
-
-impl FromIterator<bool> for BooleanChunked {
-    fn from_iter<I: IntoIterator<Item = bool>>(iter: I) -> Self {
-        // 2021-02-07: this was ~70% faster than with the builder, even with the extra Option<T> added.
-        BooleanArray::from_iter(iter.into_iter().map(Some)).into()
-    }
-}
-
-impl FromIterator<bool> for NoNull<BooleanChunked> {
-    fn from_iter<I: IntoIterator<Item = bool>>(iter: I) -> Self {
-        let ca = iter.into_iter().collect::<BooleanChunked>();
-        NoNull::new(ca)
-    }
-}
-
-// FromIterator for StringChunked variants.
-
-impl<Ptr> FromIterator<Option<Ptr>> for StringChunked
-where
-    Ptr: AsRef<str>,
-{
-    fn from_iter<I: IntoIterator<Item = Option<Ptr>>>(iter: I) -> Self {
-        let arr = MutableBinaryViewArray::from_iterator(iter.into_iter()).freeze();
-        ChunkedArray::with_chunk("", arr)
-    }
-}
-
-/// Local [`AsRef<T>`] trait to circumvent the orphan rule.
-pub trait PolarsAsRef<T: ?Sized>: AsRef<T> {}
-
-impl PolarsAsRef<str> for String {}
-impl PolarsAsRef<str> for &str {}
-// &["foo", "bar"]
-impl PolarsAsRef<str> for &&str {}
-
-impl<'a> PolarsAsRef<str> for Cow<'a, str> {}
-impl PolarsAsRef<[u8]> for Vec<u8> {}
-impl PolarsAsRef<[u8]> for &[u8] {}
-// TODO: remove!
-impl PolarsAsRef<[u8]> for &&[u8] {}
-impl<'a> PolarsAsRef<[u8]> for Cow<'a, [u8]> {}
-
-impl<Ptr> FromIterator<Ptr> for StringChunked
-where
-    Ptr: PolarsAsRef<str>,
-{
-    fn from_iter<I: IntoIterator<Item = Ptr>>(iter: I) -> Self {
-        let arr = MutableBinaryViewArray::from_values_iter(iter.into_iter()).freeze();
-        ChunkedArray::with_chunk("", arr)
-    }
-}
-
-// FromIterator for BinaryChunked variants.
-impl<Ptr> FromIterator<Option<Ptr>> for BinaryChunked
-where
-    Ptr: AsRef<[u8]>,
-{
-    fn from_iter<I: IntoIterator<Item = Option<Ptr>>>(iter: I) -> Self {
-        let arr = MutableBinaryViewArray::from_iter(iter).freeze();
-        ChunkedArray::with_chunk("", arr)
-    }
-}
-
-impl<Ptr> FromIterator<Ptr> for BinaryChunked
-where
-    Ptr: PolarsAsRef<[u8]>,
-{
-    fn from_iter<I: IntoIterator<Item = Ptr>>(iter: I) -> Self {
-        let arr = MutableBinaryViewArray::from_values_iter(iter.into_iter()).freeze();
-        ChunkedArray::with_chunk("", arr)
-    }
-}
-
-impl<Ptr> FromIterator<Ptr> for ListChunked
-where
-    Ptr: Borrow<Series>,
-{
-    fn from_iter<I: IntoIterator<Item = Ptr>>(iter: I) -> Self {
-        let mut it = iter.into_iter();
-        let capacity = get_iter_capacity(&it);
-
-        // first take one to get the dtype.
-        let v = match it.next() {
-            Some(v) => v,
-            None => return ListChunked::full_null("", 0),
-        };
-        // We don't know the needed capacity. We arbitrarily choose an average of 5 elements per series.
-        let mut builder =
-            get_list_builder(v.borrow().dtype(), capacity * 5, capacity, "collected").unwrap();
-
-        builder.append_series(v.borrow()).unwrap();
-        for s in it {
-            builder.append_series(s.borrow()).unwrap();
-        }
-        builder.finish()
-    }
-}
-
-impl FromIterator<Option<Series>> for ListChunked {
-    fn from_iter<I: IntoIterator<Item = Option<Series>>>(iter: I) -> Self {
-        let mut it = iter.into_iter();
-        let capacity = get_iter_capacity(&it);
-
-        // get first non None from iter
-        let first_value;
-        let mut init_null_count = 0;
-        loop {
-            match it.next() {
-                Some(Some(s)) => {
-                    first_value = Some(s);
-                    break;
-                },
-                Some(None) => {
-                    init_null_count += 1;
-                },
-                None => return ListChunked::full_null("", init_null_count),
-            }
-        }
-
-        match first_value {
-            None => {
-                // already returned full_null above
-                unreachable!()
-            },
-            Some(ref first_s) => {
-                // AnyValues with empty lists in python can create
-                // Series of an unknown dtype.
-                // We use the anonymousbuilder without a dtype
-                // the empty arrays is then not added (we add an extra offset instead)
-                // the next non-empty series then must have the correct dtype.
-                if matches!(first_s.dtype(), DataType::Null) && first_s.is_empty() {
-                    let mut builder = AnonymousOwnedListBuilder::new("collected", capacity, None);
-                    for _ in 0..init_null_count {
-                        builder.append_null();
-                    }
-                    builder.append_empty();
-
-                    for opt_s in it {
-                        builder.append_opt_series(opt_s.as_ref()).unwrap();
-                    }
-                    builder.finish()
-                } else {
-                    match first_s.dtype() {
-                        #[cfg(feature = "object")]
-                        DataType::Object(_, _) => {
-                            let mut builder =
-                                first_s.get_list_builder("collected", capacity * 5, capacity);
-                            for _ in 0..init_null_count {
-                                builder.append_null();
-                            }
-                            builder.append_series(first_s).unwrap();
-
-                            for opt_s in it {
-                                builder.append_opt_series(opt_s.as_ref()).unwrap();
-                            }
-                            builder.finish()
-                        },
-                        _ => {
-                            // We don't know the needed capacity. We arbitrarily choose an average of 5 elements per series.
-                            let mut builder = get_list_builder(
-                                first_s.dtype(),
-                                capacity * 5,
-                                capacity,
-                                "collected",
-                            )
-                            .unwrap();
-
-                            for _ in 0..init_null_count {
-                                builder.append_null();
-                            }
-                            builder.append_series(first_s).unwrap();
-
-                            for opt_s in it {
-                                builder.append_opt_series(opt_s.as_ref()).unwrap();
-                            }
-                            builder.finish()
-                        },
-                    }
-                }
-            },
-        }
-    }
-}
-
-impl FromIterator<Option<Box<dyn Array>>> for ListChunked {
-    fn from_iter<I: IntoIterator<Item = Option<Box<dyn Array>>>>(iter: I) -> Self {
-        let mut cap = 0;
-        let mut dtype: Option<DataType> = None;
-        let vals = iter
-            .into_iter()
-            .map(|opt_arr| {
-                opt_arr.map(|arr| {
-                    if dtype.is_none() {
-                        dtype = Some(arr.data_type().into());
-                    }
-                    cap += arr.len();
-                    arr
-                })
-            })
-            .collect::<Vec<_>>();
-
-        let mut builder = AnonymousListBuilder::new("collected", cap, None);
-        for val in &vals {
-            builder.append_opt_array(val.as_deref());
-        }
-        builder.finish()
-    }
-}
-
-#[cfg(feature = "dtype-array")]
-impl ArrayChunked {
-    /// # Safety
-    /// The caller must ensure that the underlying `Arrays` match the given datatype.
-    /// That means the logical map should map to the physical type.
-    pub unsafe fn from_iter_and_args<I: IntoIterator<Item = Option<ArrayRef>>>(
-        iter: I,
-        width: usize,
-        capacity: usize,
-        inner_dtype: Option<DataType>,
-        name: &str,
-    ) -> Self {
-        let mut builder =
-            AnonymousOwnedFixedSizeListBuilder::new(name, width, capacity, inner_dtype);
-        for val in iter {
-            match val {
-                None => builder.push_null(),
-                Some(arr) => builder.push_unchecked(arr.as_ref(), 0),
-            }
-        }
-        builder.finish()
-    }
-}
-
-#[cfg(feature = "object")]
-impl<T: PolarsObject> FromIterator<Option<T>> for ObjectChunked<T> {
-    fn from_iter<I: IntoIterator<Item = Option<T>>>(iter: I) -> Self {
-        let iter = iter.into_iter();
-        let size = iter.size_hint().0;
-        let mut null_mask_builder = MutableBitmap::with_capacity(size);
-
-        // TODO: Clippy lint is broken, remove attr once fixed.
-        // https://github.com/rust-lang/rust-clippy/issues/12580
-        #[cfg_attr(feature = "nightly", allow(clippy::manual_unwrap_or_default))]
-        let values: Vec<T> = iter
-            .map(|value| match value {
-                Some(value) => {
-                    null_mask_builder.push(true);
-                    value
-                },
-                None => {
-                    null_mask_builder.push(false);
-                    T::default()
-                },
-            })
-            .collect();
-
-        let null_bit_buffer: Option<Bitmap> = null_mask_builder.into();
-        let null_bitmap = null_bit_buffer;
-
-        let len = values.len();
-
-        let arr = Box::new(ObjectArray {
-            values: Arc::new(values),
-            null_bitmap,
-            offset: 0,
-            len,
-        });
-        let mut out = ChunkedArray {
-            field: Arc::new(Field::new("", get_object_type::<T>())),
-            chunks: vec![arr],
-            phantom: PhantomData,
-            bit_settings: Default::default(),
-            length: 0,
-            null_count: 0,
-        };
-        out.compute_len();
-        out
-    }
-}
+use crate::utils::NoNull;
 
 /// FromParallelIterator trait
 // Code taken from https://docs.rs/rayon/1.3.1/src/rayon/iter/extend.rs.html#356-366
